# openapi.json
{
 "openapi": "3.0.3",
 "info": {
  "title": "FastAPI",
  "version": "0.1.0"
 },
 "servers": [
  {
   "url": "/api"
  }
 ],
 "paths": {
  "/windows": {
   "post": {
    "summary": "Visible Windows",
    "operationId": "visible_windows_windows_post",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/WindowInfo"
         },
         "type": "array",
         "title": "Response Visible Windows Windows Post"
        }
       }
      }
     }
    }
   }
  },
  "/topmost": {
   "post": {
    "summary": "Topmost Windows",
    "operationId": "topmost_windows_topmost_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/TopMostRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/templates": {
   "get": {
    "summary": "Get All Templates",
    "operationId": "get_all_templates_templates_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/Template"
         },
         "type": "array",
         "title": "Response Get All Templates Templates Get"
        }
       }
      }
     }
    }
   },
   "post": {
    "summary": "Create Template",
    "operationId": "create_template_templates_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/TemplateContent"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "integer",
         "title": "Response Create Template Templates Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/templates/id/{id_}": {
   "get": {
    "summary": "Get Template By Id",
    "operationId": "get_template_by_id_templates_id__id___get",
    "parameters": [
     {
      "name": "id_",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Id "
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Template"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "patch": {
    "summary": "Patch Template",
    "operationId": "patch_template_templates_id__id___patch",
    "parameters": [
     {
      "name": "id_",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Id "
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/TemplateOnlyTitleOrHtml"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "type": "integer"
         },
         "title": "Response Patch Template Templates Id  Id   Patch"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/templates/title/{title}": {
   "get": {
    "summary": "Get Template By Title",
    "operationId": "get_template_by_title_templates_title__title__get",
    "parameters": [
     {
      "name": "title",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Title"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/Template"
         },
         "title": "Response Get Template By Title Templates Title  Title  Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/templates/{id_}": {
   "delete": {
    "summary": "Delete Template",
    "operationId": "delete_template_templates__id___delete",
    "parameters": [
     {
      "name": "id_",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Id "
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "type": "integer"
         },
         "title": "Response Delete Template Templates  Id   Delete"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/ping": {
   "get": {
    "summary": "Ping",
    "operationId": "ping_ping_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     }
    }
   }
  }
 },
 "components": {
  "schemas": {
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
   "Template": {
    "properties": {
     "title": {
      "type": "string",
      "title": "Title"
     },
     "html": {
      "type": "string",
      "title": "Html"
     },
     "last_modified": {
      "type": "number",
      "title": "Last Modified"
     },
     "created": {
      "type": "number",
      "title": "Created"
     },
     "id": {
      "type": "integer",
      "title": "Id"
     }
    },
    "type": "object",
    "required": [
     "title",
     "html",
     "last_modified",
     "created",
     "id"
    ],
    "title": "Template"
   },
   "TemplateContent": {
    "properties": {
     "title": {
      "type": "string",
      "title": "Title"
     },
     "html": {
      "type": "string",
      "title": "Html"
     }
    },
    "type": "object",
    "required": [
     "title",
     "html"
    ],
    "title": "TemplateContent"
   },
   "TemplateOnlyTitleOrHtml": {
    "properties": {
     "title": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Title"
     },
     "html": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Html"
     }
    },
    "type": "object",
    "title": "TemplateOnlyTitleOrHtml"
   },
   "TopMostRequest": {
    "properties": {
     "hwnd": {
      "type": "integer",
      "title": "Hwnd"
     },
     "top": {
      "type": "boolean",
      "title": "Top"
     }
    },
    "type": "object",
    "required": [
     "hwnd",
     "top"
    ],
    "title": "TopMostRequest"
   },
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
   "WindowInfo": {
    "properties": {
     "hwnd": {
      "type": "integer",
      "title": "Hwnd"
     },
     "title": {
      "type": "string",
      "title": "Title"
     },
     "is_top": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Is Top"
     },
     "name_of_pid": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Name Of Pid"
     },
     "exe_of_pid": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Exe Of Pid"
     }
    },
    "type": "object",
    "required": [
     "hwnd",
     "title",
     "is_top"
    ],
    "title": "WindowInfo"
   }
  }
 }
}